first workin prototype of fast_steering
