+++ src/bioemu/config/steering/physical_steering.yaml
@@ -1,21 +1,21 @@
 do_steering: true
-num_particles: 3
+num_particles: 5
 start: 0.5
-end: 1.0 # End time for steering (default: continue until end)
-resample_every_n_steps: 5
+end: 0.9 # End time for steering (default: continue until end)
+resample_every_n_steps: 1
 fast_steering: true
 potentials:
   chainbreak:
     _target_: bioemu.steering.ChainBreakPotential
-    tolerance: 1.
-    slope: 1.
-    max_value: 100
-    order: 1
-    linear_from: 1.
+    tolerance: 0.5
+    slope: 5.
+    max_value: 10000
+    order: 2
+    linear_from: 10
     weight: 1.0
   chainclash:
     _target_: bioemu.steering.ChainClashPotential
     tolerance: 0.
     dist: 4.1
-    slope: 3.
+    slope: 5.
     weight: 1.0
