+++ src/bioemu/shortcuts.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # Quick way to refer to things to instantiate in the config
-from .denoiser import dpm_solver, heun_denoiser, euler_maruyama_denoiser  # noqa
+from .denoiser import dpm_solver, heun_denoiser # noqa
 from .models import DiGConditionalScoreModel  # noqa
 from .sde_lib import CosineVPSDE  # noqa
 from .so3_sde import DiGSO3SDE  # noqa
